Added bin information for npm link so you can utilize:
node-sass sass.scss output.css

Where sass.scss is the Sass file to compilte, and output.css is the filename to output.

Added fs.writeFile functionality to the node-sass bin, so that output if read in and compiled correct are then output to the filesystem based on the given arguments. Default output file is "nodesass.scss"

Did not add additional tests for this, lame.