Fix automake tests when sass-spec-dir is absolute.
It seems that automake gets confused with an absolute test path, even if
that path is contained in this directory. It doesn't seem to care if the
directory is outside the source directory, so we just make the absolute
paths relative. Not so easy in sh, but fortunately, we already require
ruby in this region of configure.