Fix formatting of unmodified colors
Colors which have no operations applied to them are normally not
converted to a standard output.  This commit tracks the original color
format and outputs it if no changes have occurred, matching compass
output.

It does this in a slightly roundabout way, mostly so that when
https://github.com/nex3/sass/issues/363 is done on the sass side, a few
lines can be removed and the original format will be kept much more
widely to match.