inspect: always round color channels before inspection
Since RGB channels are stored as doubles, HSL/RGB conversions can result
in decimal values being stored for a channel. This breaks color name
lookup. For example, previously

    rgb(127.5, 127.5, 127.5)

would fail to be converted to "gray," despite being rounded and output
as

    #808080

which *does* map to gray. So instead of performing the rounding only
when outputting as a hex triplet, perform the rounding at the start,
always. This achieves consistency with Sass.