Removal of option hashed based success and error callbacks This change removes the object hash success/error callbacks in favor of node style callbacks. This includes README and test updates, including the removal of two tests that are now obsolete.
This also updates the CLI code pathway to use the node style
callback.

BREAKING API CHANGE
This removes the success/error option keys as ways to register
callbacks into node-sass. Instead, a function should be passed as
the second parameter for the asyncronous `render` call. It takes
a standard node signature of `error, callback`.