Improve memory manager for MSVC exceptions in c-tor args
MSVC calls allocator first before evaluating arguments for
the actual object constructor. Since the evaluation phase
can quite easily throw an error (like in `parse_list`), the
memory manager is left in an undefined state. The object is
already tracked but not initialized and we get an error if
we try to `delete` the object (we should use free instead).

Added macro to be used instead of overloaded new operator!