Use process.config.target_defaults.default_configuration as the build type default
node-gyp is using process.config.target_defaults.default_configuration
to check which configuration is about to be used, but this
can be overriden by particular engine's config.gypi file.

Since we do not support such a cross compilation anyway,
use this variable in place of a hardcoded default.

Instead of introducing another option to force the use
of 'Release' we should read and pass options to node-gyp.
This is not implemented here.