Should throw too many arguments error with varargs
This PR fixes a missing error when expanding varargs results in
too many arguments for a function or mixin. It also addresses
some inconsistencies in the warning messages. These inconsistencies
are inconsistent with previous inconsistencies also. I've opened
a bug with the Ruby Sass team to help slow out these rough edges.
https://github.com/sass/sass/issues/1880