Don't allocate std::random_device statically
It appears that on Linux with gcc (tested with gcc 4.6.3 on Ubuntu 12.04 and
gcc 5.2.1 on Ubuntu 15.10) std::random_device will open /dev/urandom, and keep
the file open for the lifetime of the std::random_device object.

Since libsass allocates the std::random_device globally, it will not be
destroyed if the processes execs. Futhermore std::random_device does not set
close-on-exec on the fd, so the fd will be leaked to the child process.

The std::random_device object is only used by the GetSeed function, and as the
GetSeed function is only called once to seed the rand variable, there is no
need to keep the std::random_device object.

So this patch moves the std::random_device object into the GetSeed function,
thereby eliminating the fd-leak.