Fix import resolution order changing with custom importers
When using the following syntax with a custom importer

```css
@import "foo", "bar";
```

Paths that are matched by a custom importer get pushed into the
queue ahead of those that don't. This means if `bar` is matched
by a custom importer the import order will be changed to
`"bar", "foo"` instead of the intended `"foo", "bar"`. This can
easily break user code.

See sass/node-sass#1381 for an example