Remove vendor directory on prepublish and pack This is necessary because if a user is bundling node-sass in their package, the vendor directory should stay so that the user isn't required to run 'npm rebuild' afterwards, assuming that it was originally installed on a like-architecture.
However, at the same time, when publishing, we don't want to include the
binaries since they may not work for a user plus there's the build step
that tries to download the correct binary for the current architecture.

Unfortunately, npm's prepublish hook runs on both prepublish *and* npm
install (see npm/npm#10074), so, using in-publish as a workaround to get
the prepublish script to only run prepublish and pack but not on
install.

Fixes #1183
Closes #1384