Don't use import resource cache if there are custom importers
The import resource cache prevent custom importers from being
called for all imports.

This happens because we during the import resolution we cache
the resources parsed AST tree and simply substitute it for future
reference to that resource.

If that resource has it's own imports those imports are only
resolved the first time the file is parsed. Although this is a
nice optimisation, it does negatively affect custom importers.

Rather than disabling the resource caching wholesale, this patch
disables it only if a custom importer has been registered.