Improve error handling in C-API to be more robust
It's possible that another throw happens in error handling.
One case is invalid utf8 in error messages that json wants
to stringify. We could wire it up to handle arbitrary level
of nested throws, but that could lead to endless loops, so
I opted to only wire it for one additional level.