Replace deprecated npmconf package.
This was previously attempted in #1413. Shortly after it's
release proxy users started experiencing installation issues so
this was reverted. It was later determined that #1458 was likely
at fault for the proxy issues.

Full credit for this patch goes to @delitescere.

I've also taken the liberty of cleaning the request config generation.

Fixes #1333