Error when declarations are printed without ruleset
This is introduces a vastly reduced `CheckNesting` visitor from Ruby
Sass as discussed in #2061. Doing this properly will require some
small changes to the parser, and probably eval, which currently try
to do some nesting checking of their own.

This is just first step to properly introducing the proper nesting
checking.

Closes #2061
Fixes #1732