Consolidate the Declaration and Propset AST nodes
In order to fix #1585 we need to introduce the `Trace` node from
Ruby Sass. This node interacts with propsets in an interesting way.
Have two node to represent the single Ruby `Prop` node caused a
lot of trouble with correct implementing check nesting logic for
`Trace`.

The trouble with implementing `Trace` were exaggerated by us
incorrectly eval propsets into declarations in the expand step.
This PR moves the de-nesting of prop nodes the cssize step to
[better match Ruby Sass][1].

Later we should rename this node `Prop` but it's out of scope of
this PR. This should allow my work on porting the `Trace` node over
from Ruby Sass to fix #1585.

[1]: https://github.com/sass/sass/blob/63586f1e51c0aec47a7afcc8cd17aa03f32f0a29/lib/sass/tree/visitors/cssize.rb#L149-L166