Add the Trace AST node
Extracted from #2086. This PR adds the `Trace` AST node but doesn't
implement any of it's usage. This change is a noop. It's intended
simply to reduce the noise in #2086.

The Trace node is a simple wrapper around the result of a mixin call
and `@content`. It exists as a way to group the resulting properties
as a block but maintain some metadata about the initiating mixin call
or `@content` block.