Add a timeout to the binary fetch request
This PR adds a timeout to the binary fetch request as well as a
useful error message.

It came to light in #1690 that the download can block for a long time.

```
npm info lifecycle node-sass@3.8.0~install: node-sass@3.8.0

> node-sass@3.8.0 install /Usersproject-path/node_modules/node-sass
> node scripts/install.js

Build timed out (after 35 minutes). Marking the build as failed.
```

We see a fair amount of hung installer issues that could be related
to this. The error output will give us another data point in those
cases.