Retry CLI specs that fail three times
For some reason our test fail due to timeouts on seemingly random
cli tests. This instability has been growing worse. I've spent the
better part of the last two days looking into the issue without a
resolution. I'm starting to think it's an artefact cause by the
virtualisation used in CI, potentially related to resources
starvation.

The specs that fail do so trying to spawn a child process. The
process never spawns eventually resulting in a timeout. This is why
@nschonni attempt at addressing this by increasing the timeout in

Since the spawn failure is random, and catastrophic the best solution
I can think of in the short-medium term is to instruct mocha to retry
failing tests individually a couple times.

I've scoped this behaviour to the cli specs because at the moment
we have some specs that are not reentrant.