Increase the binary fetch request timeout
I found the 1s timeout to be really fragile when testing locally.
The intent of the timeout is to prevent the install process hanging
indefinitely. It makes sense to be very generous so this increases
the timeout to 60s.

The timeout applies to both the initial connection as well as
individual packets.