Remove skip download for CI flag
This flag means we skip a bunch of code paths in CI which is not
ideal. Since we tell people to use `npm rebuild` we should eat our
own dog food.

With this patch we do an install, then do a rebuild hitting all the
major install code paths.

Fixes #1453