Change sort order for source files (sort by compile time)
This should give slightly faster build times when
compiling in parallel with many threads.

bigger files (in terms of compile time) tend to go to the top,
so they don't end up as the last compile unit when compiling
in parallel. But we also want to mix them a little too avoid
heavy RAM usage peaks. Other than that the order is arbitrary.