Handle permission errors when attempting to cache binaries
This uses a candidate model for determining an suitable cache
directory. This is different from the current approach with eagerly
chooses a cache directory. If that directory isn't accessible for
some reason (i.e. permissions) we have to bail out. With this
approach we can try a prioritised list of candidate directories.

Additionally this address the case where npm cache config is
blank.

Fixes #1781