Fix incorrect validation of mixin definition
Removed the duplicate validation check from the parser. The validation
in the check nesting visitor was being skipped.

Looks like there is a bunch of left of AST validation in the parser,
which now belongs in the check nesting visitor. With the current
approach the parser should only ever error when it encounters invalid
tokens. It cannot accurately validate the AST because things like
`at-root` change the structure in important ways.

Closes #2269
Fixes #2095
Spec sass/sass-spec#1031