Remove Sass_List_Delimiter in favour of an is_bracketed property
The list delimiter property only existed as a way to preserve parser
state on the node in order to reused the existing `parse_list` function.
This PR instead adds a `parse_bracketed_list` function to explicitly
parse bracketed lists and sets the `is_bracketed` property on the `List`
node. It turns out this is exactly what Ruby Sass does.