Restore behaviour of cwd in include path
Traditionally Sass has added the cwd as the first include path. This
behaviour was remove in Sass 3.4, and recently in LibSass 3.5.0.beta.2.
People depend on this behaviour and as a result a lot of compilations
are now failing.

This PR restores the expected behaviour and adds a test to boot.

Fixes #1876