Fix watching of CSS files to avoid infinite build cycle
Previously when the src and dest dirs were the same, whenever a CSS file
changed (either manually or as the result of a build) it was recursively
recompiled in an infinite loop, unless a build finished in less than
500ms (Gaze's default file watching debounce). Now, when the input and
output dirs are the same, CSS files are treated like includes (files
starting with an underscore) in that they are never built into CSS, but
they are watched in case they are included in a Sass file.

When the input and output dirs differ, functionality is unchanged.