Fix memory leak and internal leak detector
Media_Blocks are prone to have circular references.
Copy could leak memory if it does not get picked up.
Looks like we are able to reset block reference for copy
Good as it will ensure a low memory overhead for this fix
So this is a cheap solution with a minimal price