Add global install instructions, change phrasing
I have added the way to globally install node-sass from the CLI,
which isn't present in the README otherwise. I have also removed a
sentence which says that node-sass is simplistic. In both cases,
it is assumed that someone using the program knows how to use npm
to install command line packages, and that the user has some idea
about the level of complexity usual to an npm package. Both of
these assumptions are assumptions - my wording makes the package
more friendly to beginners by not assuming they know this.

tl;dr: just a small change to help new users.