Pre-empt support for Node 10
Lets try to avoid the rush of issues when Node 10 is released next
month.

Node 10 is currently version [`62`][1] but is expected to be
[`63`][2] before going stable. We need to pick one because of the
ABI breakage betweem 62 and 63.

If this bet pays off we save ourselves a bunch of time dealing with
installation issues, and avoiding a new release. If we're wrong it's
no different to any other major Node release.

[1]: https://github.com/nodejs/node/blob/97595739/src/node_version.h#L112
[2]: https://github.com/nodejs/node/pull/19201