Clarify docs for --source-map. Closes #1026.
It is not clear from the `--help` docs that `--source-map` requires a boolean or path. Passing the flag alone results in a TypeError. There's an argument to be made that passing the flag alone should imply `true`, or at the very least it should fail with a helpful error, not a cryptic TypeError. For now, this PR fixes the --help docs so that they match the README.