Bump node-gyp; disable binary builds in CI
Technically this is a breaking change for Node < 4. We still see
a bananas amount of downloads for Node 0.10 on CI - presumably CI
environments.

The breakage is limited only to building the binaries locally. This
means node-sass will continue to install, and operate as normal on
all our support Node 0.10 and 0.12 environments.

IMHO this is a fair trade off to address the `npm audit` noise without
the disruption of forcing a major version bump before we're ready.